requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
twilio==8.14.0
google-cloud-speech==2.24.1 